requirements.txt
opencv-python==4.7.0.72
opencv-contrib-python==4.7.0.72
pyzmq==25.1.0
setproctitle==1.3.2
py-trees==2.2.3